pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["3FANG <d.hasimikov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
uvicorn = "^0.27.1"
django-bootstrap5 = "^23.4"
django-filter = "^24.2"
rollbar = "^1.0.0"
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
